requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
docopt==0.6.2
future==0.16.0
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
lxml
networkx==2.1
nltk==3.2.5
numpy==1.14.2
parso==0.2.0
Penman==0.6.2
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
PyDelphin==0.6.2
Pygments==2.2.0
PyYAML==3.12
requests==2.18.4
sacrebleu
simplegeneric==0.8.1
six==1.11.0
torch==0.3.1
torchtext==0.2.3
tqdm==4.23.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
